refactor: Ajustei o favorit button para ficar em cima da imagem no canto inferior direito, e fiz a as importações de forma correta para aplicar os estilos
