+++ src/pages/home/components/BlackFridaySection.jsx
