+++ src/pages/home/components/FavoritesDropdown.jsx
