+++ src/pages/home/components/ProductInfoSection.jsx
