+++ src/styles/css/product_card.css
@@ -246,6 +246,54 @@ html, body, #root {
 }
 
 .product-card {
+  position: relative;
+}
+.product-card .favorite-button-wrapper {
+  position: absolute;
+  bottom: 208px;
+  right: 15px;
+  z-index: 2;
+  opacity: 0;
+  visibility: hidden;
+  transition: all 0.3s ease;
+  transform: scale(0.8);
+}
+.product-card:hover .favorite-button-wrapper {
+  opacity: 1;
+  visibility: visible;
+  transform: scale(1);
+}
+
+.favorite-button {
+  width: 36px;
+  height: 36px;
+  border-radius: 50%;
+  background: white;
+  border: none;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  transition: all 0.2s ease;
+  box-shadow: 0 2px 4px rgba(0, 51, 33, 0.1);
+}
+.favorite-button i {
+  font-size: 1.2rem;
+  color: #666;
+  transition: all 0.2s ease;
+}
+.favorite-button:hover {
+  transform: scale(1.1);
+  box-shadow: 0 3px 6px rgba(0, 51, 33, 0.15);
+}
+.favorite-button:hover i {
+  color: #ff4b4b;
+}
+.favorite-button.active i {
+  color: #ff4b4b;
+}
+
+.product-card {
   transition: transform 0.2s ease-in-out;
   background-color: #fff;
   border: 1px solid #eee;
@@ -285,6 +333,7 @@ html, body, #root {
   display: -webkit-box;
   -webkit-line-clamp: 2;
   -webkit-box-orient: vertical;
+  line-clamp: 2;
 }
 
 .fs-product-card-price {
