+++ src/styles/sass/_favorite_button.scss
@@ -1,10 +1,10 @@
 .product-card {
   position: relative;
-  
+
   .favorite-button-wrapper {
     position: absolute;
-    top: 10px;
-    right: 10px;
+    bottom: 208px;
+    right: 15px;
     z-index: 2;
     opacity: 0;
     visibility: hidden;
@@ -54,4 +54,4 @@
       color: #ff4b4b;
     }
   }
-} 
+}
