+++ src/styles/sass/_favorites.scss
