+++ src/styles/sass/_favorites_dropdown.scss
